deps.qt: Add QtSerialPort
The obs-ptz plugin requires the QtSerialPort library, but it isn't
currently built by obs-deps. obs-ptz is carrying it's own build of
qtserialport, but this solution isn't portable. It causes library
mismatches resulting in crashes when obs-studio changes the version of
Qt being used. This is because the qtserialport library itself uses the
Qt internal API which isn't stable between versions, and so needs to be
the same version as the rest of the Qt libraries.

The easiest solution is to build qtserialport in obs-deps, and then ship
the qtserialport library with obs-studio so that plugins don't need to
try and match library versions. However, there has been resistance to
this solution in the past because obs-studio itself doesn't need
qtserialport and there was concern over the build cost of additional
libraries. I'm hoping that qtserialport is small enough that it will be
reconsidered for inclusion.

This patch simply adds qtserialport to the obs-deps build.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>